--- app/src/main/java/ca/judacribz/week2weekend/global/constants/WebConstants.kt
@@ -1,8 +1,0 @@
-package ca.judacribz.week2weekend.global.constants
-
-const val TAG_H2 = "h2"
-const val TAG_P = "p"
-const val TAG_A = "a"
-
-const val ATTR_STYLE = "style"
-const val ATTR_HREF = "href"
