--- app/src/main/java/ca/judacribz/week2weekend/global/viewmodel/BaseViewModel.kt
@@ -1,20 +1,0 @@
-package ca.judacribz.week2weekend.global.viewmodel
-
-import androidx.lifecycle.ViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlin.coroutines.CoroutineContext
-
-open class BaseViewModel : ViewModel(), CoroutineScope {
-
-    private val job by lazyOf(Job())
-
-    override fun onCleared() {
-        super.onCleared()
-        job.cancel()
-    }
-
-    override val coroutineContext: CoroutineContext
-        get() = Dispatchers.Default + job
-}
