--- app/src/main/java/ca/judacribz/week2weekend/homepage/model/AnimalPost.kt
@@ -1,10 +1,0 @@
-package ca.judacribz.week2weekend.homepage.model
-
-import android.graphics.Bitmap
-
-data class AnimalPost(
-    val image: Bitmap,
-    val headline: String?,
-    val shortDescription: String?,
-    val learnMoreUrl: String?
-)
