+++ app/src/main/java/ca/judacribz/zooatlanta/global/constants/ZooPageConstants.kt
@@ -1,0 +1,11 @@
+package ca.judacribz.zooatlanta.global.constants
+
+const val ZOO_ATLANTA_URL = "https://zooatlanta.org/"
+
+// To retrieve schedule
+const val ID_TODAY = "today"
+const val ID_HOURS_TODAY = "hours-today"
+
+// To retrieve homepage images and text
+const val CLASS_SLIDE = "slide"
+const val CLASS_HERO_IMAGE = "hero-image"
