+++ app/src/main/java/ca/judacribz/zooatlanta/global/viewmodel/BaseViewModel.kt
@@ -1,0 +1,20 @@
+package ca.judacribz.zooatlanta.global.viewmodel
+
+import androidx.lifecycle.ViewModel
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlin.coroutines.CoroutineContext
+
+open class BaseViewModel : ViewModel(), CoroutineScope {
+
+    private val job by lazyOf(Job())
+
+    override fun onCleared() {
+        super.onCleared()
+        job.cancel()
+    }
+
+    override val coroutineContext: CoroutineContext
+        get() = Dispatchers.Default + job
+}
