+++ app/src/main/java/ca/judacribz/zooatlanta/homepage/model/AnimalPost.kt
@@ -1,0 +1,10 @@
+package ca.judacribz.zooatlanta.homepage.model
+
+import android.graphics.Bitmap
+
+data class AnimalPost(
+    val image: Bitmap,
+    val headline: String?,
+    val shortDescription: String?,
+    val learnMoreUrl: String?
+)
