+++ app/src/main/java/ca/judacribz/zooatlanta/homepage/model/Schedule.kt
@@ -1,0 +1,3 @@
+package ca.judacribz.zooatlanta.homepage.model
+
+data class Schedule(val admissionTime: String, val lastAdmission: String)
