+++ app/src/main/res/layout/activity_homepage.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:background="@drawable/bg"
+    android:background="@color/colorZooAtlanta"
     android:id="@+id/clHomepage"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
@@ -50,6 +50,7 @@
         android:id="@+id/tvSchedule"
         android:layout_height="wrap_content"
         android:layout_width="0dp"
+        android:textColor="@android:color/black"
         android:textSize="17sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toEndOf="@id/tvToday"
@@ -62,6 +63,7 @@
         android:id="@+id/tvLastAdmin"
         android:layout_height="wrap_content"
         android:layout_width="0dp"
+        android:textColor="@android:color/black"
         android:textSize="17sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toEndOf="@id/tvToday"
@@ -117,6 +119,7 @@
 
     <LinearLayout
         android:background="@android:color/black"
+        android:gravity="center"
         android:id="@+id/homepage_ll_learn_more"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
